+++ src/components/organisms/TodoListComponent.tsx
@@ -1,4 +1,4 @@
-import { formatToYYYY_MM_DD } from '@/helpers/dateUtils';
+import { formatToYYYY_MM_DD } from '@/helpers/dateHelper';
 import { useReadScheduleQuery } from '@/redux/schedule/slice';
 import { TodoCardComponent } from '../atoms/TodoCardComponent';
